+++ demos/assets-1m.c
@@ -153,7 +153,6 @@
 #include "../res/assets/default/inc/images/arrow_left_n.res"
 #include "../res/assets/default/inc/images/middle_on.res"
 #include "../res/assets/default/inc/images/bricks.res"
-#include "../res/assets/default/inc/images/msg.res"
 #include "../res/assets/default/inc/images/play_o.res"
 #include "../res/assets/default/inc/images/play_n.res"
 #include "../res/assets/default/inc/images/msg_active.res"
@@ -166,7 +165,6 @@
 #include "../res/assets/default/inc/images/check.res"
 #include "../res/assets/default/inc/images/cursor.res"
 #include "../res/assets/default/inc/images/me.res"
-#include "../res/assets/default/inc/images/info.res"
 #include "../res/assets/default/inc/images/red_btn_p.res"
 #include "../res/assets/default/inc/images/middle_off.res"
 #include "../res/assets/default/inc/images/arrow_down_n.res"
@@ -190,7 +188,6 @@
 #include "../res/assets/default/inc/images/red_btn_n.res"
 #include "../res/assets/default/inc/images/arrow_down_p.res"
 #include "../res/assets/default/inc/images/slider_v_bg.res"
-#include "../res/assets/default/inc/images/warn.res"
 #include "../res/assets/default/inc/images/green_btn_p.res"
 #include "../res/assets/default/inc/images/switch.res"
 #include "../res/assets/default/inc/images/radio_unchecked.res"
@@ -310,8 +307,6 @@
 #include "../res/assets/default/inc/images/slider_drag_p.data"
 #include "../res/assets/default/inc/images/slider_v_fg.data"
 #include "../res/assets/default/inc/images/arrow_down_n.data"
-#include "../res/assets/default/inc/images/info.data"
-#include "../res/assets/default/inc/images/msg.data"
 #include "../res/assets/default/inc/images/num_7.data"
 #include "../res/assets/default/inc/images/arrow_left_n.data"
 #include "../res/assets/default/inc/images/edit_clear_p.data"
@@ -493,7 +488,6 @@ ret_t assets_init(void) {
   assets_manager_add(am, image_num_0);
   assets_manager_add(am, image_arrow_down_p);
   assets_manager_add(am, image_num_1);
-  assets_manager_add(am, image_warn);
   assets_manager_add(am, image_contact_active);
   assets_manager_add(am, image_slider_drag_o);
   assets_manager_add(am, image_battery_4);
@@ -520,8 +514,6 @@ ret_t assets_init(void) {
   assets_manager_add(am, image_slider_drag_p);
   assets_manager_add(am, image_slider_v_fg);
   assets_manager_add(am, image_arrow_down_n);
-  assets_manager_add(am, image_info);
-  assets_manager_add(am, image_msg);
   assets_manager_add(am, image_num_7);
   assets_manager_add(am, image_arrow_left_n);
   assets_manager_add(am, image_edit_clear_p);