+++ src/fscript_ext/fscript_iostream_serial.c
@@ -21,14 +21,47 @@
 
 static ret_t func_iostream_serial_create(fscript_t* fscript, fscript_args_t* args,
                                          value_t* result) {
-  int baudrate = 0;
+  int baudrate = 115200;
   object_t* obj = NULL;
   const char* devname = NULL;
-
-  result->type = VALUE_TYPE_INVALID;
+  bytesize_t bytesize = eightbits;
+  parity_t parity = parity_none;
+  stopbits_t stopbits = stopbits_one;
+  flowcontrol_t flowcontrol = flowcontrol_none;
   FSCRIPT_FUNC_CHECK(args->size >= 1, RET_BAD_PARAMS);
+
   devname = value_str(args->args);
-  baudrate = args->size > 1 ? value_int(args->args + 1) : 0;
+  if (args->size > 1) {
+    baudrate = value_int(args->args + 1);
+  }
+  if (args->size > 2) {
+    bytesize = value_int(args->args + 2);
+  }
+  if (args->size > 3) {
+    const char* str = value_str(args->args + 3);
+    if (str != NULL) {
+      if (strstr(str, "odd") != NULL) {
+        parity = parity_odd;
+      } else if (strstr(str, "even") != NULL) {
+        parity = parity_even;
+      }
+    }
+  }
+  if (args->size > 4) {
+    stopbits = value_int(args->args + 4);
+  }
+  if (args->size > 5) {
+    const char* str = value_str(args->args + 5);
+    if (str != NULL) {
+      if (strstr(str, "hard") != NULL) {
+        flowcontrol = flowcontrol_hardware;
+        parity = parity_odd;
+      } else if (strstr(str, "soft") != NULL) {
+        flowcontrol = flowcontrol_software;
+      }
+    }
+  }
+
   return_value_if_fail(devname != NULL, RET_BAD_PARAMS);
   obj = OBJECT(tk_iostream_serial_create(devname));
   return_value_if_fail(obj != NULL, RET_BAD_PARAMS);
@@ -36,10 +69,7 @@ static ret_t func_iostream_serial_create(fscript_t* fscript, fscript_args_t* arg
   value_set_object(result, obj);
   result->free_handle = TRUE;
 
-  if (baudrate > 0) {
-    object_set_prop_int(OBJECT(obj), TK_IOSTREAM_SERIAL_PROP_BAUDRATE, baudrate);
-    object_exec(OBJECT(obj), TK_IOSTREAM_SERIAL_CMD_CONFIG, NULL);
-  }
+  tk_iostream_serial_config(obj, baudrate, bytesize, parity, stopbits, flowcontrol);
 
   return RET_OK;
 }