+++ src/widgets/pages.c
@@ -31,11 +31,15 @@ static bool_t pages_target_is_page(widget_t* target) {
   return target->vt != NULL && tk_str_eq(target->vt->type, WIDGET_TYPE_PAGES);
 }
 
+static bool_t pages_active_is_valid(widget_t* widget, uint32_t active) {
+  return (active < widget_count_children(widget));
+}
+
 static ret_t pages_save_target(widget_t* widget) {
   pages_t* pages = PAGES(widget);
   return_value_if_fail(pages != NULL, RET_BAD_PARAMS);
 
-  if (widget->children != NULL && widget->children->size > pages->active) {
+  if (pages_active_is_valid(widget, pages->active)) {
     widget_t* active_view = widget_get_child(widget, pages->active);
 
     if (active_view != NULL) {
@@ -62,7 +66,10 @@ static ret_t pages_restore_target(widget_t* widget) {
   widget_t* active_view = NULL;
   pages_t* pages = PAGES(widget);
   return_value_if_fail(pages != NULL, RET_BAD_PARAMS);
-  active_view = widget_get_child(widget, pages->active);
+
+  if (pages_active_is_valid(widget, pages->active)) {
+    active_view = widget_get_child(widget, pages->active);
+  }
 
   if (active_view != NULL) {
     target = default_focused_child_get_save_target(widget, active_view);
@@ -135,7 +142,7 @@ static widget_t* pages_find_target(widget_t* widget, xy_t x, xy_t y) {
   pages_t* pages = PAGES(widget);
   return_value_if_fail(pages != NULL, NULL);
 
-  if (widget->children == NULL) {
+  if (!pages_active_is_valid(widget, pages->active)) {
     return NULL;
   }
 
@@ -165,12 +172,19 @@ static ret_t pages_set_prop(widget_t* widget, const char* name, const value_t* v
 }
 
 static ret_t pages_on_event(widget_t* widget, event_t* e) {
+  pages_t* pages = PAGES(widget);
   uint16_t type = e->type;
-  return_value_if_fail(widget != NULL && e != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(widget != NULL && pages != NULL && e != NULL, RET_BAD_PARAMS);
 
   switch (type) {
-    case EVT_WINDOW_WILL_OPEN: {
-      if (PAGES(widget)->active >= widget_count_children(widget)) {
+    case EVT_WIDGET_ADD_CHILD:
+    case EVT_WIDGET_REMOVE_CHILD: {
+      if (widget->loading) {
+        break;
+      }
+    }
+    case EVT_WIDGET_LOAD: {
+      if (!pages_active_is_valid(widget, pages->active)) {
         pages_set_active(widget, 0);
       } else {
         pages_show_active(widget);
@@ -213,6 +227,24 @@ static ret_t pages_on_destroy(widget_t* widget) {
   return RET_OK;
 }
 
+static ret_t pages_on_remove_child(widget_t* widget, widget_t* child) {
+  int32_t active = -1;
+  int32_t remove_index = -1;
+  pages_t* pages = PAGES(widget);
+  return_value_if_fail(widget != NULL && pages != NULL && child != NULL, RET_BAD_PARAMS);
+
+  remove_index = widget_index_of(child);
+  return_value_if_fail(remove_index >= 0, RET_BAD_PARAMS);
+
+  active = (int32_t)(pages->active);
+  if (remove_index < active ||
+      (remove_index == active && remove_index == widget->children->size - 1)) {
+    active = tk_max(active - 1, 0);
+    pages->active = (uint32_t)active;
+  }
+  return RET_CONTINUE;
+}
+
 static const char* const s_pages_clone_properties[] = {WIDGET_PROP_VALUE, NULL};
 
 TK_DECL_VTABLE(pages) = {.size = sizeof(pages_t),
@@ -228,6 +260,7 @@ TK_DECL_VTABLE(pages) = {.size = sizeof(pages_t),
                          .on_event = pages_on_event,
                          .get_prop = pages_get_prop,
                          .set_prop = pages_set_prop,
+                         .on_remove_child = pages_on_remove_child,
                          .on_destroy = pages_on_destroy};
 
 widget_t* pages_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h) {