+++ src/widgets/tab_button_group.c
@@ -138,31 +138,60 @@ static ret_t tab_button_group_on_paint_children(widget_t* widget, canvas_t* c) {
   return hscrollable_on_paint_children(tab_button_group->hscrollable, c);
 }
 
-static ret_t tab_button_group_ensure_active(widget_t* widget) {
-  widget_t* first = NULL;
+static ret_t tab_button_group_update_active(widget_t* widget) {
+  tab_button_group_t* tab_button_group = TAB_BUTTON_GROUP(widget);
+  return_value_if_fail(tab_button_group != NULL, RET_BAD_PARAMS);
 
   WIDGET_FOR_EACH_CHILD_BEGIN(widget, iter, i)
   if (widget_get_value(iter)) {
+    tab_button_group->active = i;
     return RET_OK;
-  }
-  if (first == NULL) {
-    first = iter;
   }
   WIDGET_FOR_EACH_CHILD_END();
 
+  tab_button_group->active = -1;
+  return RET_OK;
+}
+
+static ret_t tab_button_group_ensure_active(widget_t* widget) {
+  widget_t* first = NULL;
+  tab_button_group_t* tab_button_group = TAB_BUTTON_GROUP(widget);
+  return_value_if_fail(tab_button_group != NULL, RET_BAD_PARAMS);
+
+  tab_button_group_update_active(widget);
+
+  if (-1 != tab_button_group->active) {
+    return RET_OK;
+  }
+
+  first = widget_get_child(widget, 0);
   if (first != NULL) {
+    tab_button_group->active = 0;
     widget_set_value(first, TRUE);
   }
 
   return RET_OK;
 }
 
+static bool_t tab_button_group_active_is_valid(widget_t* widget, uint32_t active) {
+  return (active < widget_count_children(widget));
+}
+
 static ret_t tab_button_group_on_event(widget_t* widget, event_t* e) {
   tab_button_group_t* tab_button_group = TAB_BUTTON_GROUP(widget);
   return_value_if_fail(tab_button_group != NULL, RET_BAD_PARAMS);
 
-  if (e->type == EVT_WINDOW_WILL_OPEN) {
-    tab_button_group_ensure_active(widget);
+  switch (e->type) {
+    case EVT_WINDOW_WILL_OPEN: {
+      tab_button_group_ensure_active(widget);
+      break;
+    }
+    case EVT_WIDGET_REMOVE_CHILD: {
+      if (tab_button_group_active_is_valid(widget, tab_button_group->active)) {
+        widget_set_value(widget_get_child(widget, tab_button_group->active), TRUE);
+      }
+      break;
+    }
   }
 
   if (tab_button_group->scrollable && tab_button_group->compact) {
@@ -206,6 +235,29 @@ static ret_t tab_button_group_get_offset(widget_t* widget, xy_t* out_x, xy_t* ou
   return RET_OK;
 }
 
+static ret_t tab_button_group_on_remove_child(widget_t* widget, widget_t* child) {
+  int32_t remove_index = -1;
+  tab_button_group_t* tab_button_group = TAB_BUTTON_GROUP(widget);
+  return_value_if_fail(
+      widget != NULL && widget->children != NULL && tab_button_group != NULL && child != NULL,
+      RET_BAD_PARAMS);
+
+  remove_index = widget_index_of(child);
+  return_value_if_fail(remove_index >= 0, RET_BAD_PARAMS);
+
+  tab_button_group_update_active(widget);
+
+  if (-1 != tab_button_group->active) {
+    uint32_t children_num = widget->children->size;
+    if (remove_index < tab_button_group->active ||
+        (remove_index == tab_button_group->active && remove_index == children_num - 1)) {
+      tab_button_group->active--;
+    }
+  }
+
+  return RET_CONTINUE;
+}
+
 TK_DECL_VTABLE(tab_button_group) = {.size = sizeof(tab_button_group_t),
                                     .type = WIDGET_TYPE_TAB_BUTTON_GROUP,
                                     .scrollable = TRUE,
@@ -216,6 +268,7 @@ TK_DECL_VTABLE(tab_button_group) = {.size = sizeof(tab_button_group_t),
                                     .on_event = tab_button_group_on_event,
                                     .on_destroy = tab_button_group_on_destroy,
                                     .get_offset = tab_button_group_get_offset,
+                                    .on_remove_child = tab_button_group_on_remove_child,
                                     .on_paint_children = tab_button_group_on_paint_children,
                                     .on_layout_children = tab_button_group_on_layout_children};